# Tune HTTP client for concurrency