Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>

<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->

Tune HTTP client for concurrency

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
- [ ] I have raised an issue to propose this change ([required](https://github.com/openfaas/faas/blob/master/CONTRIBUTING.md))

- due to what appears to be a frequent issue with the Go HTTP
client some tweaks were needed to the HTTP client used for
reverse proxying to prevent CoreDNS from rejecting connections.

The following PRs / commits implement similar changes in
Prometheus and Minio.

https://github.com/prometheus/prometheus/pull/3592
https://github.com/minio/minio/pull/5860

These tweaks can be overridden with an environmental variable:

* `max_idle_conns` (default 1024)
* `max_idle_conns_per_host` (default 1024)

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

Under a 3-node (1-master) kubeadm cluster running on bare
metal with Ubuntu 18.04 I was able to send 100k requests
with 1000 being concurrent with no errors being returned
by hey.

```
helm upgrade openfaas --install openfaas/openfaas  \
   --namespace openfaas     \
   --set basic_auth=true  \
   --set functionNamespace=openfaas-fn \
  --set operator.create=true \
  --set gateway.image=openfaas/gateway:0.9.15-rc5d \
  --set gateway.scale_from_zero=true
```

I tested with and without adding the override to `gateway.image=`

On each machine in the cluster and the sender I ran:

```
sudo sysctl -w net.core.somaxconn=1024
```

> Note the nginx ingress controller config: recommends 65k for this value.

[hey](https://github.com/rakyll/hey)

```
hey -n 100000 -c 1000 -m=POST -d="hi" \
  http://192.168.0.26:31112/function/go-echo
```

The go-echo function is based upon the golang-http
template in the function store using the of-watchdog.

Update your gateway to use the following image: `openfaas/gateway:0.9.15-rc4`.

## Remaining work

* Compare & contrast with `scale_from_zero` enabled / disabled

Scale from zero does add some latency to some of the requests, especially in parallel

* Test image on Docker Swarm

I did a few basic tests as above with a single node, no errors with a single node.

* Make the max_conn* values configurable

This has been done in the latest commits

* Find out what the resulting state of the client machine is via `ss -ta` after a large concurrent batch of requests - `ulimit -n` may be needed. My default was `1024` open files.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I've read the [CONTRIBUTION](https://github.com/openfaas/faas/blob/master/CONTRIBUTING.md) guide
- [x] I have signed-off my commits with `git commit -s`
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.