@alexellis 2019-01-27
A tag with the latest commits and the new environment variables will be published under `openfaas/gateway:5e78156833dce5d07ec2bf6b454232c5ea309acf`