@LucasRoesler 2019-02-01
@alexellis I deployed this on GKE using Helm and the following values file

```yaml
basic_auth: true
functionNamespace: openfaas-fn
exposeServices: false
ingress:
    enabled: true
    hosts:
      - host: $GATEWAY
        path: /
        serviceName: gateway
        servicePort: 8080
gateway:
    image: openfaas/gateway:5e78156833dce5d07ec2bf6b454232c5ea309acf
```
```sh
helm upgrade openfaas --install openfaas/openfaas \
    --debug \
    --namespace openfaas \
    --values $HOME/Code/of-concurrency-test.yaml
```

I then deployed the `NodeInfo` function and load tested it with `hey`

```sh
$ export GATEWAY=of.dev.contiamo.io
$ faas-cli store deploy NodeInfo --gateway http://$GATEWAY
$ hey -n 1000 -c 10 -m=POST http://$GATEWAY/function/nodeinfo -a admin:$PASSWORD

[truncated...]

Status code distribution:
  [200]	1000 responses

$ hey -n 5000 -c 50 -m=POST http://$GATEWAY/function/nodeinfo -a admin:$PASSWORD

[truncated...]

Status code distribution:
  [200]	4934 responses
  [502]	30 responses

Error distribution:
  [15]	Post http://of.dev.contiamo.io/function/nodeinfo: EOF
  [21]	Post http://of.dev.contiamo.io/function/nodeinfo: http: server closed idle connect
```