@alexellis 2019-02-01
Hi Lucas,

Thanks so much for looking into this.

Can you not use node info, but the test function I gave?

NodeInfo is using forking so it may hit the file-descriptor limit of the Pod.

Alex