@alexellis 2019-02-01
For the test function: https://github.com/alexellis/echo-fn