@LucasRoesler 2019-02-01
Ok, after deploying `echo-fn`

```sh
git clone https://github.com/alexellis/echo-fn
cd echo-fn
faas-cli template store pull golang-http
faas-cli deploy --gateway http://$GATEWAY

hey -n 5000 -c 50 -m=POST -d="hi there" http://$GATEWAY/function/go-echo -a admin:$PASSWORD

[truncated...]

Status code distribution:
  [200]	5000 responses
```