@LucasRoesler 2019-02-01
And pushing it a little bit further

```sh
hey -n 10000 -c 50 -m=POST -d="hi there" http://$GATEWAY/function/go-echo -a admin:$PASSWORD

[truncated...]

Status code distribution:
  [200]	10000 responses
```