@alexellis 2019-02-02
@stefanprodan 
@LucasRoesler 

Please could you both follow-up on your original reviews now that I've completed the changes and we've done some testing.

BT via @noppen also confirmed this fixed their benchmarking issue