@LucasRoesler 2019-02-02
@alexellis  I just re-ran the test on GKE using the echo-fn that you linked to.


I used helm with the following values
```sh
cat > "of-concurrency-test.yaml" <<- EndOfMessage
basic_auth: true
functionNamespace: openfaas-fn
exposeServices: false
ingress:
    enabled: true
    hosts:
      - host: $GATEWAY
        path: /
        serviceName: gateway
        servicePort: 8080
gateway:
    image: openfaas/gateway:5e78156833dce5d07ec2bf6b454232c5ea309acf
    scale_from_zero: true
EndOfMessage

helm upgrade openfaas --install openfaas/openfaas \
    --debug \
    --namespace openfaas \
    --values of-concurrency-test.yaml
```

I then deployed alexellis/echo-fn@02c8acb and then tested three times, working up to your desired test load

```
$ hey -n 1000 -c 10 -m=POST -d="hi there" http://$GATEWAY/function/go-echo
Status code distribution:
  [200]	1000 responses

$ hey -n 10000 -c 100 -m=POST -d="hi there" http://$GATEWAY/function/go-echo
Status code distribution:
  [200]	10000 responses

$ hey -n 100000 -c 1000 -m=POST -d="hi there" http://$GATEWAY/function/go-echo
Status code distribution:
  [200]	89158 responses
  [404]	2 responses
  [504]	1703 responses
```

Unfortunately, I tried to chase down relevant logs for those 504 and 404 responses but none of the containers seem to have any interesting logs (i checked the ingress controller, the gateway, and faas-netes) Nginx does log the 504 responses and there was nothing in the logs from the OF components.