@alexellis 2019-02-02
This is many times better than the previous configuration. There are two new environment variables you can tune which may help, as well as somaxconn on the nodes and "loader" machine.