@alexellis 2019-02-02
There may also be some forms of rate limiting done by GCP on LBs.