@alexellis 2019-02-02
We might also want to try a higher replica count with min/max counts being the same to exclude scaling from any stats.