@alexellis 2019-02-02
I'm pretty sure I am getting rate-limited using `n1-standard-2 (2 vCPUs, 7.5 GB memory)`

I went up to 6 nodes, 10 gateways and 50 static replicas of the function then did the load test from a Packet.net host (250GB RAM, 48 CPUS.. I know.. ) `ulimit -n 60000` and `sysctl -w net.core.somaxconn=65000`.

No ingress was used, just an LB and this config for the gateway:

```
 65         - name: max_idle_conns
 66           value: "20000"
 67         - name: max_idle_conns_per_host
 68           value: "1024"
```

```
root@load:~# go/bin/hey -n 10000 -c 1000 -m="POST" -d="hi" http://35.240.4.250:8080/function/go-echo

Summary:
  Total:	22.1521 secs
  Slowest:	16.7710 secs
  Fastest:	0.0112 secs
  Average:	0.7105 secs
  Requests/sec:	451.4248
  
  Total data:	20000 bytes
  Size/request:	2 bytes

Response time histogram:
  0.011 [1]	|
  1.687 [9501]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.363 [341]	|■
  5.039 [28]	|
  6.715 [25]	|
  8.391 [28]	|
  10.067 [24]	|
  11.743 [15]	|
  13.419 [17]	|
  15.095 [11]	|
  16.771 [9]	|


Latency distribution:
  10% in 0.0410 secs
  25% in 0.1573 secs
  50% in 0.5279 secs
  75% in 0.8585 secs
  90% in 1.3024 secs
  95% in 1.6785 secs
  99% in 6.9973 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0060 secs, 0.0112 secs, 16.7710 secs
  DNS-lookup:	0.0000 secs, 0.0000 secs, 0.0000 secs
  req write:	0.0017 secs, 0.0000 secs, 0.0700 secs
  resp wait:	0.7004 secs, 0.0111 secs, 16.7709 secs
  resp read:	0.0001 secs, 0.0000 secs, 0.0241 secs

Status code distribution:
  [200]	10000 responses
```