@alexellis 2019-02-02
2nd run:

```
root@load:~# go/bin/hey -n 10000 -c 1000 -m="POST" -d="hi" http://35.240.4.250:8080/function/go-echo

Summary:
  Total:	21.9627 secs
  Slowest:	16.7202 secs
  Fastest:	0.0111 secs
  Average:	0.4853 secs
  Requests/sec:	455.3171
  
  Total data:	20000 bytes
  Size/request:	2 bytes

Response time histogram:
  0.011 [1]	|
  1.682 [9526]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.353 [328]	|■
  5.024 [40]	|
  6.695 [19]	|
  8.366 [16]	|
  10.037 [18]	|
  11.707 [16]	|
  13.378 [18]	|
  15.049 [9]	|
  16.720 [9]	|


Latency distribution:
  10% in 0.0205 secs
  25% in 0.0597 secs
  50% in 0.2011 secs
  75% in 0.5122 secs
  90% in 1.0409 secs
  95% in 1.5968 secs
  99% in 5.4021 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0031 secs, 0.0111 secs, 16.7202 secs
  DNS-lookup:	0.0000 secs, 0.0000 secs, 0.0000 secs
  req write:	0.0012 secs, 0.0000 secs, 0.0767 secs
  resp wait:	0.4750 secs, 0.0109 secs, 16.7200 secs
  resp read:	0.0001 secs, 0.0000 secs, 0.0147 secs

Status code distribution:
  [200]	10000 responses
```