@noppen 2019-02-04
No worries, see our testing details below.

# OpenFaas Benchmarking

### Environment setup

**Host Specifications**  
- Intel(R) Xeon(R) CPU E5-2620 v4 32 Cores @ 2.10GHz  
- 64GB Memory  

In each instance, invalid requests are sent to our endpoint. This as the aim of this test was not to test the performance of the function, but rather a high volume of requests aimed at the infrastructure sitting in front of it. As such 400 responses from our endpoint sufficed and should be considered successful in the report below.

The following load generation command was used in each instance:

```bash
hey -n 100000 -c 1500 -m=POST -d="" <url>
```

## Tests

### Old OpenFaas Gateway

#### With net.core.somaxconn=128

```
Summary:
  Total:        104.3133 secs
  Slowest:      19.9578 secs
  Fastest:      0.0012 secs
  Average:      1.1717 secs
  Requests/sec: 949.0639

  Total data:   2650153 bytes
  Size/request: 27 bytes

Response time histogram:
  0.001 [1]     |
  1.997 [80682] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.993 [13609] |■■■■■■■
  5.988 [2715]  |■
  7.984 [669]   |
  9.980 [73]    |
  11.975 [25]   |
  13.971 [78]   |
  15.967 [70]   |
  17.962 [16]   |
  19.958 [22]   |


Latency distribution:
  10% in 0.1557 secs
  25% in 0.3356 secs
  50% in 0.6975 secs
  75% in 1.5162 secs
  90% in 2.8530 secs
  95% in 3.7412 secs
  99% in 5.9758 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0029 secs, 0.0012 secs, 19.9578 secs
  DNS-lookup:   0.0014 secs, 0.0000 secs, 0.1329 secs
  req write:    0.0003 secs, 0.0000 secs, 0.0657 secs
  resp wait:    1.1684 secs, 0.0012 secs, 19.9578 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0539 secs

Status code distribution:
  [200] 82661 responses
  [404] 305 responses
  [500] 7152 responses
  [502] 7842 responses

Error distribution:
  [1040]        Post <SNIP>: net/http: request canceled (Client.Timeout exceeded while awaiting headers)
```

### Old OpenFaas Gateway

#### With net.core.somaxconn=65000

```
Summary:
  Total:        98.8088 secs
  Slowest:      19.9633 secs
  Fastest:      0.0014 secs
  Average:      1.1219 secs
  Requests/sec: 1001.9352

  Total data:   2693558 bytes
  Size/request: 27 bytes

Response time histogram:
  0.001 [1]     |
  1.998 [80724] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.994 [11153] |■■■■■■
  5.990 [4710]  |■■
  7.986 [865]   |
  9.982 [102]   |
  11.979 [16]   |
  13.975 [25]   |
  15.971 [18]   |
  17.967 [25]   |
  19.963 [17]   |


Latency distribution:
  10% in 0.1732 secs
  25% in 0.3044 secs
  50% in 0.5753 secs
  75% in 1.2173 secs
  90% in 3.1690 secs
  95% in 4.2016 secs
  99% in 6.1667 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0054 secs, 0.0014 secs, 19.9633 secs
  DNS-lookup:   0.0031 secs, 0.0000 secs, 0.2580 secs
  req write:    0.0010 secs, 0.0000 secs, 0.1497 secs
  resp wait:    1.1151 secs, 0.0013 secs, 19.9632 secs
  resp read:    0.0001 secs, 0.0000 secs, 0.1615 secs

Status code distribution:
  [200] 77054 responses
  [404] 199 responses
  [500] 14798 responses
  [502] 5605 responses

Error distribution:
  [1344]        Post <SNIP>: net/http: request canceled (Client.Timeout exceeded while awaiting headers)
```

### New OpenFaas Gateway

#### With net.core.somaxconn=128

```
Summary:
  Total:        25.6389 secs
  Slowest:      15.7911 secs
  Fastest:      0.0028 secs
  Average:      0.3695 secs
  Requests/sec: 3861.3178

  Total data:   2871000 bytes
  Size/request: 29 bytes

Response time histogram:
  0.003 [1]     |
  1.582 [98427] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.160 [443]   |
  4.739 [30]    |
  6.318 [15]    |
  7.897 [15]    |
  9.476 [14]    |
  11.055 [13]   |
  12.633 [26]   |
  14.212 [10]   |
  15.791 [6]    |


Latency distribution:
  10% in 0.2922 secs
  25% in 0.3190 secs
  50% in 0.3492 secs
  75% in 0.3838 secs
  90% in 0.4286 secs
  95% in 0.4641 secs
  99% in 1.2127 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0027 secs, 0.0028 secs, 15.7911 secs
  DNS-lookup:   0.0013 secs, 0.0000 secs, 0.1236 secs
  req write:    0.0003 secs, 0.0000 secs, 0.0711 secs
  resp wait:    0.3664 secs, 0.0027 secs, 15.6029 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0908 secs

Status code distribution:
  [400] 99000 responses
```

### New OpenFaas Gateway

#### With net.core.somaxconn=65000

```
Summary:
  Total:        25.3714 secs
  Slowest:      18.2375 secs
  Fastest:      0.0017 secs
  Average:      0.3584 secs
  Requests/sec: 3902.0248

  Total data:   2870971 bytes
  Size/request: 29 bytes

Response time histogram:
  0.002 [1]     |
  1.825 [98308] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.649 [224]   |
  5.472 [400]   |
  7.296 [19]    |
  9.120 [18]    |
  10.943 [14]   |
  12.767 [9]    |
  14.590 [5]    |
  16.414 [0]    |
  18.237 [1]    |


Latency distribution:
  10% in 0.2298 secs
  25% in 0.2937 secs
  50% in 0.3371 secs
  75% in 0.3773 secs
  90% in 0.4213 secs
  95% in 0.4426 secs
  99% in 1.5349 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0024 secs, 0.0017 secs, 18.2375 secs
  DNS-lookup:   0.0002 secs, 0.0000 secs, 0.1825 secs
  req write:    0.0001 secs, 0.0000 secs, 0.0517 secs
  resp wait:    0.3558 secs, 0.0017 secs, 18.0536 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.1112 secs

Status code distribution:
  [400] 98999 responses

Error distribution:
  [1]   Post <SNIP>: net/http: request canceled (Client.Timeout exceeded while awaiting headers)
```

### Without OpenFaas Gateway

#### With net.core.somaxconn=128

```
Summary:
  Total:        24.5840 secs
  Slowest:      18.6884 secs
  Fastest:      0.0050 secs
  Average:      0.3102 secs
  Requests/sec: 4027.0105

  Total data:   2871000 bytes
  Size/request: 29 bytes

Response time histogram:
  0.005 [1]     |
  1.873 [98134] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.742 [179]   |
  5.610 [242]   |
  7.478 [185]   |
  9.347 [53]    |
  11.215 [0]    |
  13.083 [1]    |
  14.952 [13]   |
  16.820 [0]    |
  18.688 [192]  |


Latency distribution:
  10% in 0.1011 secs
  25% in 0.1978 secs
  50% in 0.2665 secs
  75% in 0.3011 secs
  90% in 0.3236 secs
  95% in 0.3390 secs
  99% in 0.9469 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0030 secs, 0.0050 secs, 18.6884 secs
  DNS-lookup:   0.0018 secs, 0.0000 secs, 0.1403 secs
  req write:    0.0005 secs, 0.0000 secs, 0.1206 secs
  resp wait:    0.3066 secs, 0.0049 secs, 18.3740 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.1373 secs

Status code distribution:
  [400] 99000 responses
```

### Without OpenFaas Gateway

#### With net.core.somaxconn=65000

```
Summary:
  Total:        24.5417 secs
  Slowest:      13.2278 secs
  Fastest:      0.0015 secs
  Average:      0.3394 secs
  Requests/sec: 4033.9574

  Total data:   2871000 bytes
  Size/request: 29 bytes

Response time histogram:
  0.002 [1]     |
  1.324 [98789] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  2.647 [0]     |
  3.969 [0]     |
  5.292 [0]     |
  6.615 [0]     |
  7.937 [0]     |
  9.260 [0]     |
  10.583 [0]    |
  11.905 [0]    |
  13.228 [210]  |


Latency distribution:
  10% in 0.2327 secs
  25% in 0.2883 secs
  50% in 0.3223 secs
  75% in 0.3559 secs
  90% in 0.3875 secs
  95% in 0.4301 secs
  99% in 0.5857 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0024 secs, 0.0015 secs, 13.2278 secs
  DNS-lookup:   0.0002 secs, 0.0000 secs, 0.1858 secs
  req write:    0.0001 secs, 0.0000 secs, 0.0308 secs
  resp wait:    0.3369 secs, 0.0015 secs, 12.9997 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0768 secs

Status code distribution:
  [400] 99000 responses
```